change browser.c dates
I don't actually know what I'm doing here, but it works enough for
me and I don't really care...More importantly, this now leaves
another 10 characters of space for the filename, which is good on
my small tablet screen as I can actually see enough of the
filenames to be useful now.

Note: for the given use of the tool, I don't get why the timestamp
is so long. I think it would be better to do timestamps the way exa
does, instead of a standard format. show time if same day; date if
same year; else year/month; this gives enough information in
general. either way, the timezone is almost certainly unnecessary

This repo will now be archived and likely won't be updated, though
I am uploading a release (mainly for myself, tho...). If you don't
want to deal with the linker warning when using it, theres probably
patches in termux-packages to fix that.